pyproject.toml
[tool.poetry]
name = "openpowerlifting"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"


[tool.poetry.group.dev.dependencies]
psycopg2-binary = "2.9.9"
pandas = "2.2.2"
transformers = "4.41.0"
torch = "2.3.0"
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
